text_extraction/requirements.txt
PyPDF2==1.26.0
pytesseract==0.3.8
Pillow==8.2.0
